Create typography comp folder and move bodytext and heading
