+++ ngx-fudis/projects/ngx-fudis/src/lib/components/typography/body-text/body-text.component.scss
@@ -1,4 +1,4 @@
-@use '../../foundations/typography/mixins' as typography;
+@use '../../../foundations/typography/mixins' as typography;
 
 .fudis-body-text {
 	&__s-regular {
