+++ ngx-fudis/projects/ngx-fudis/src/lib/components/typography/heading/heading.component.scss
@@ -1,4 +1,4 @@
-@import '../../foundations/typography/mixins';
+@import '../../../foundations/typography/mixins';
 
 .fudis-heading {
 	&__xxs {
