+++ ngx-fudis/projects/ngx-fudis/src/lib/ngx-fudis.module.ts
@@ -12,8 +12,8 @@ import {
 } from './components/dialog/dialog-directives';
 import { FudisDialog } from './components/dialog/dialog';
 import { ButtonComponent } from './components/button/button.component';
-import { HeadingComponent } from './components/heading/heading.component';
-import { BodyTextComponent } from './components/body-text/body-text.component';
+import { HeadingComponent } from './components/typography/heading/heading.component';
+import { BodyTextComponent } from './components/typography/body-text/body-text.component';
 import { TextInputComponent } from './components/form/text-input/text-input.component';
 import { VanillaTextInputComponent } from './components/form/vanilla-text-input/vanilla-text-input.component';
 import { FormLayoutComponent } from './components/form/form-layout/form-layout.component';
