+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -4,8 +4,8 @@
 
 export * from './lib/ngx-fudis.module';
 export * from './lib/components/dialog/public-api';
-export { HeadingComponent } from './lib/components/heading/heading.component';
-export { BodyTextComponent } from './lib/components/body-text/body-text.component';
+export { HeadingComponent } from './lib/components/typography/heading/heading.component';
+export { BodyTextComponent } from './lib/components/typography/body-text/body-text.component';
 export { ButtonComponent } from './lib/components/button/button.component';
 export { TextInputComponent } from './lib/components/form/text-input/text-input.component';
 export { VanillaTextInputComponent } from './lib/components/form/vanilla-text-input/vanilla-text-input.component';
